BNA modifications  - Changed alpha from float to double  - Updated ThingKey clone functions to be more automatic  - Changed key data parameters to group like keys together  - Removed shape thing  - Formatting corrections for *ThingBase.java files  - Fixed: missing readOnly attribute for ImplementsFacet

+++ org.archstudio.bna/src/org/archstudio/bna/facets/IHasAlpha.java
@@ -18,8 +18,8 @@ import org.eclipse.jdt.annotation.Nullable;
 @NonNullByDefault
 public interface IHasAlpha extends org.archstudio.bna.IThing {
 
-	public static final IThingKey<java.lang.Float> ALPHA_KEY = ThingKey.create(com.google.common.collect.Lists.newArrayList("alpha", IHasAlpha.class));
+	public static final IThingKey<java.lang.Double> ALPHA_KEY = ThingKey.create(com.google.common.collect.Lists.newArrayList("alpha", IHasAlpha.class));
 
-	public float getAlpha();
+	public double getAlpha();
 
 }
