+++ org.archstudio.bna/src/org/archstudio/bna/facets/IHasGlow.java
@@ -18,13 +18,13 @@ import org.eclipse.jdt.annotation.Nullable;
 @NonNullByDefault
 public interface IHasGlow extends org.archstudio.bna.IThing {
 
-	public static final IThingKey<java.lang.Float> GLOW_ALPHA_KEY = ThingKey.create(com.google.common.collect.Lists.newArrayList("glowAlpha", IHasGlow.class));
+	public static final IThingKey<java.lang.Double> GLOW_ALPHA_KEY = ThingKey.create(com.google.common.collect.Lists.newArrayList("glowAlpha", IHasGlow.class));
 
 	public static final IThingKey<org.eclipse.swt.graphics.RGB> GLOW_COLOR_KEY = ThingKey.create(com.google.common.collect.Lists.newArrayList("glowColor", IHasGlow.class), true);
 
 	public static final IThingKey<java.lang.Integer> GLOW_WIDTH_KEY = ThingKey.create(com.google.common.collect.Lists.newArrayList("glowWidth", IHasGlow.class));
 
-	public float getGlowAlpha();
+	public double getGlowAlpha();
 
 	public @Nullable org.eclipse.swt.graphics.RGB getGlowColor();
