+++ org.archstudio.bna/src/org/archstudio/bna/ui/swt/SWTResources.java
@@ -143,7 +143,7 @@ public class SWTResources extends AbstractUIResources implements ISWTResources {
 				continue;
 			}
 			try {
-				pushAlpha((float)thingToRender.get(IHasAlpha.ALPHA_KEY, 1f));
+				pushAlpha(thingToRender.get(IHasAlpha.ALPHA_KEY, 1d));
 				try {
 					//gc.setTint(thingToRender.get(IHasTint.TINT_KEY, new RGB(0, 0, 0)));
 					IThingPeer<?> peer = view.getThingPeer(thingToRender);
